feat: Refactor GenericInfo and Update UI Components
This commit refactors the `GenericInfo` interface and updates UI components to better align with KMP (Kotlin Multiplatform) principles. It removes Android-specific dependencies and improves code consistency.

- **GenericInfo Refactoring:**
    - Moved the `GenericInfo` interface to `kmpuiviews` to make it platform-agnostic.
    - Updated `GenericInfo` to not have a `context`, `activity`, or `fragmentActivity` parameter, and removed the need for them.
    - Updated the `chapterOnClick` and `downloadChapter` to no longer need `context` or `fragmentActivity` or `activity`.
    - Updated the `debugMenuItem` to no longer need a context.
    - Added `KmpGenericInfo` to `kmpuiviews` that has all of the functions that `GenericInfo` has.
    - Created a `MockInfo` that takes a context.
- **UI Component Updates:**
    - Updated `GlobalSearchFragment` to use `kmpuiviews` `ComponentState` instead of the `uiviews` version.
    - Moved `ComposeSettingsDsl` and `ComponentState` to `kmpuiviews`.
    - Updated `OtakuCustomListScreen` and `CoverCard` to use `kmpuiviews` `ComponentState`.
    - Updated `DetailsFragment` to use the updated `chapterOnClick` and `downloadChapter` functions.
    - Updated `BaseMainActivity` to use `kmpuiviews` `ComposeSettingsDsl`.
    - Updated `NotificationFragment` to use the updated `MockInfo` with context.
    - Removed usages of the old `ComposeSettingsDsl`.
    - Updated `GenericNovel`, `GenericAnime`, and `GenericManga` to use the updated `chapterOnClick` and `downloadChapter` functions.
    - Updated `GenericAnime` to launch a coroutine to get chapter info instead of using `activity.lifecycleScope.launch`.
    - Updated `GenericAnime` to use `downloadVideo` and `getEpisodes` correctly and not inside an if statement.
    - Added `ComponentState` to `kmpuiviews`.
- **Other Changes:**
   - Removed `toSource` and `sourceList` from `GenericInfo`.
   - Removed `ApiServiceDeserializer` dependency on `genericInfo.toSource`.
   - Updated `Screen.kt` to remove the `genericInfo` dependency.
   - Updated `OnboardingScreen` to use `kmpuiviews` `ComposeSettingsDsl`.
   - Removed `fragmentActivity` and `activity` parameters from multiple functions.
   - Removed `uiviews` `ComponentState` enum class.
   - Updated `composeCustomPreferences` in `GenericManga` and `GenericAnime` to use the correct way of using the functions.
   - Removed `sourceList`, `searchList`, and `toSource` from `GenericAnime`, `GenericManga`, and `GenericNovel` since it was no longer needed.
   - Created `KmpGenericInfo` to have the new functions.
