+++ UIViews/src/main/java/com/programmersbox/uiviews/GenericInfo.kt
@@ -4,42 +4,20 @@ import android.annotation.SuppressLint
 import android.app.PendingIntent
 import android.content.Context
 import android.net.Uri
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.lazy.LazyItemScope
-import androidx.compose.foundation.lazy.grid.LazyGridState
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.core.net.toUri
-import androidx.fragment.app.FragmentActivity
-import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavType
 import androidx.navigation.serialization.generateRouteWithArgs
-import com.programmersbox.favoritesdatabase.DbModel
-import com.programmersbox.kmpmodels.KmpApiService
-import com.programmersbox.kmpmodels.KmpChapterModel
-import com.programmersbox.kmpmodels.KmpInfoModel
 import com.programmersbox.kmpmodels.KmpItemModel
 import com.programmersbox.kmpuiviews.presentation.Screen
 import com.programmersbox.sharedutils.AppUpdate
-import com.programmersbox.uiviews.presentation.settings.ComposeSettingsDsl
-import com.programmersbox.uiviews.utils.ComponentState
+import com.programmersbox.kmpuiviews.GenericInfo as KmpGenericInfo
 
-interface GenericInfo {
+interface GenericInfo : KmpGenericInfo {
 
     val apkString: AppUpdate.AppUpdates.() -> String?
-    val scrollBuffer: Int get() = 2
-    val deepLinkUri: String
 
-    val sourceType: String get() = ""
-
-    //TODO: Would need to be different
-    // Probably pass a context into the constructor?
     fun deepLinkDetails(context: Context, itemModel: KmpItemModel?): PendingIntent?
 
-    //TODO: Would need to be different
     fun deepLinkSettings(context: Context): PendingIntent?
 
     @SuppressLint("RestrictedApi")
@@ -66,83 +44,4 @@ interface GenericInfo {
     }
 
     fun deepLinkSettingsUri() = "$deepLinkUri${Screen.NotificationScreen.route}".toUri()
-
-    //TODO: Would need to be different
-    fun chapterOnClick(
-        model: KmpChapterModel,
-        allChapters: List<KmpChapterModel>,
-        infoModel: KmpInfoModel,
-        context: Context,
-        activity: FragmentActivity,
-        navController: NavController,
-    )
-
-    fun sourceList(): List<KmpApiService>
-    fun searchList(): List<KmpApiService> = sourceList()
-
-    //TODO: Can be removed
-    fun toSource(s: String): KmpApiService?
-    fun composeCustomPreferences(): ComposeSettingsDsl.() -> Unit = {}
-
-    //TODO: Would need to be different
-    fun downloadChapter(
-        model: KmpChapterModel,
-        allChapters: List<KmpChapterModel>,
-        infoModel: KmpInfoModel,
-        context: Context,
-        activity: FragmentActivity,
-        navController: NavController,
-    )
-
-    @Composable
-    fun DetailActions(infoModel: KmpInfoModel, tint: Color) = Unit
-
-    @Composable
-    fun ComposeShimmerItem()
-
-    @ExperimentalFoundationApi
-    @Composable
-    fun ItemListView(
-        list: List<KmpItemModel>,
-        favorites: List<DbModel>,
-        listState: LazyGridState,
-        onLongPress: (KmpItemModel, ComponentState) -> Unit,
-        modifier: Modifier,
-        paddingValues: PaddingValues,
-        onClick: (KmpItemModel) -> Unit,
-    )
-
-    @ExperimentalFoundationApi
-    @Composable
-    fun AllListView(
-        list: List<KmpItemModel>,
-        favorites: List<DbModel>,
-        listState: LazyGridState,
-        onLongPress: (KmpItemModel, ComponentState) -> Unit,
-        modifier: Modifier,
-        paddingValues: PaddingValues,
-        onClick: (KmpItemModel) -> Unit,
-    ) = ItemListView(list, favorites, listState, onLongPress, modifier, paddingValues, onClick)
-
-    @ExperimentalFoundationApi
-    @Composable
-    fun SearchListView(
-        list: List<KmpItemModel>,
-        favorites: List<DbModel>,
-        listState: LazyGridState,
-        onLongPress: (KmpItemModel, ComponentState) -> Unit,
-        modifier: Modifier,
-        paddingValues: PaddingValues,
-        onClick: (KmpItemModel) -> Unit,
-    ) = ItemListView(list, favorites, listState, onLongPress, modifier, paddingValues, onClick)
-
-    //TODO: Would need to be different
-    fun debugMenuItem(context: Context): List<@Composable LazyItemScope.() -> Unit> = emptyList()
-
-    fun NavGraphBuilder.globalNavSetup() = Unit
-
-    fun NavGraphBuilder.settingsNavSetup(): Unit = Unit
-
-    @Composable
-    fun DialogSetups() = Unit
 }
