+++ UIViews/src/main/java/com/programmersbox/uiviews/presentation/onboarding/OnboardingScreen.kt
@@ -32,9 +32,9 @@ import com.programmersbox.kmpuiviews.presentation.components.NormalOtakuScaffold
 import com.programmersbox.kmpuiviews.presentation.onboarding.Onboarding
 import com.programmersbox.kmpuiviews.presentation.onboarding.OnboardingIndicator
 import com.programmersbox.kmpuiviews.presentation.onboarding.rememberOnboardingScope
+import com.programmersbox.kmpuiviews.utils.ComposeSettingsDsl
 import com.programmersbox.sharedutils.AppLogo
 import com.programmersbox.uiviews.R
-import com.programmersbox.uiviews.presentation.settings.ComposeSettingsDsl
 import com.programmersbox.uiviews.utils.HideNavBarWhileOnScreen
 import kotlinx.coroutines.launch
 import org.koin.compose.koinInject
