+++ kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/GenericInfo.kt
@@ -1,0 +1,90 @@
+package com.programmersbox.kmpuiviews
+
+import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.lazy.LazyItemScope
+import androidx.compose.foundation.lazy.grid.LazyGridState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import com.programmersbox.favoritesdatabase.DbModel
+import com.programmersbox.kmpmodels.KmpChapterModel
+import com.programmersbox.kmpmodels.KmpInfoModel
+import com.programmersbox.kmpmodels.KmpItemModel
+import com.programmersbox.kmpuiviews.utils.ComponentState
+import com.programmersbox.kmpuiviews.utils.ComposeSettingsDsl
+
+interface GenericInfo {
+    val scrollBuffer: Int get() = 2
+    val deepLinkUri: String
+    val sourceType: String get() = ""
+
+    fun chapterOnClick(
+        model: KmpChapterModel,
+        allChapters: List<KmpChapterModel>,
+        infoModel: KmpInfoModel,
+        navController: NavController,
+    )
+
+    fun composeCustomPreferences(): ComposeSettingsDsl.() -> Unit = {}
+
+    fun downloadChapter(
+        model: KmpChapterModel,
+        allChapters: List<KmpChapterModel>,
+        infoModel: KmpInfoModel,
+        navController: NavController,
+    )
+
+    @Composable
+    fun DetailActions(infoModel: KmpInfoModel, tint: Color) = Unit
+
+    @Composable
+    fun ComposeShimmerItem()
+
+    @ExperimentalFoundationApi
+    @Composable
+    fun ItemListView(
+        list: List<KmpItemModel>,
+        favorites: List<DbModel>,
+        listState: LazyGridState,
+        onLongPress: (KmpItemModel, ComponentState) -> Unit,
+        modifier: Modifier,
+        paddingValues: PaddingValues,
+        onClick: (KmpItemModel) -> Unit,
+    )
+
+    @ExperimentalFoundationApi
+    @Composable
+    fun AllListView(
+        list: List<KmpItemModel>,
+        favorites: List<DbModel>,
+        listState: LazyGridState,
+        onLongPress: (KmpItemModel, ComponentState) -> Unit,
+        modifier: Modifier,
+        paddingValues: PaddingValues,
+        onClick: (KmpItemModel) -> Unit,
+    ) = ItemListView(list, favorites, listState, onLongPress, modifier, paddingValues, onClick)
+
+    @ExperimentalFoundationApi
+    @Composable
+    fun SearchListView(
+        list: List<KmpItemModel>,
+        favorites: List<DbModel>,
+        listState: LazyGridState,
+        onLongPress: (KmpItemModel, ComponentState) -> Unit,
+        modifier: Modifier,
+        paddingValues: PaddingValues,
+        onClick: (KmpItemModel) -> Unit,
+    ) = ItemListView(list, favorites, listState, onLongPress, modifier, paddingValues, onClick)
+
+    fun debugMenuItem(): List<@Composable LazyItemScope.() -> Unit> = emptyList()
+
+    fun NavGraphBuilder.globalNavSetup() = Unit
+
+    fun NavGraphBuilder.settingsNavSetup(): Unit = Unit
+
+    @Composable
+    fun DialogSetups() = Unit
+}
