+++ kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/utils/ComposeSettingsDsl.kt
@@ -1,0 +1,29 @@
+package com.programmersbox.kmpuiviews.utils
+
+import androidx.compose.runtime.Composable
+import com.programmersbox.kmpuiviews.presentation.onboarding.OnboardingScope
+
+class ComposeSettingsDsl {
+    //TODO: Turn back to internal once settings move to kmpuiviews
+    var generalSettings: @Composable () -> Unit = {}
+    var viewSettings: @Composable () -> Unit = {}
+    var playerSettings: @Composable () -> Unit = {}
+
+    var onboardingSettings: OnboardingScope.() -> Unit = {}
+
+    fun generalSettings(block: @Composable () -> Unit) {
+        generalSettings = block
+    }
+
+    fun viewSettings(block: @Composable () -> Unit) {
+        viewSettings = block
+    }
+
+    fun playerSettings(block: @Composable () -> Unit) {
+        playerSettings = block
+    }
+
+    fun onboardingSettings(block: OnboardingScope.() -> Unit) {
+        onboardingSettings = block
+    }
+}
